refactored app
