--- backend/fernec/models.py
@@ -1,11 +1,0 @@
-from pydantic import BaseModel
-
-class ImageItem(BaseModel):
-    image_base64: str
-
-class ImagePrediction(BaseModel):
-    predictions: list[float]
-    emotion: str
-
-class VideoPrediction(BaseModel):
-    prediction: list[str]
